ASoC: max98090: use pclk field to compare with pclk clocks
When MAX98090 is setup as I2S master there is a necessary
to setup Mi and Ni value. fPCLK value is prescaled
frequency from fMCLK by value 1, 2 or 4 depending on
clock value. For instance when fMCLK=24MHz the prescaler
value is 2 and in a result fPCLK=12MHz.

From that reason it's not wise to compare both clocks
together. Current code only works good for situation
where (accroding to spec) fMCLK is lower or equal 20MHz
because only in that situation prescaler configuration
value is 1 and fMCLK and fPCLK have the same values.

Fixing by compare valid values with valid fields.